requirements.txt
opencv-python~=4.10.0.84
numpy<2.0.0
fastapi~=0.111.0
pillow~=10.3.0
ultralytics~=8.2.46